Handles network connection for services n homepage trailer if null/empty

+++ app/src/main/java/com/example/movieappcompose/model/response/Movie.kt
@@ -14,7 +14,7 @@ data class Movie(
     val adult: Boolean? = false,
     val backdrop_path: String? = "",
     val genre_ids: List<Int> = listOf(),
-    val id: Int = 0,
+    val id: Long = 0,
     val original_language: String? = "",
     val original_title: String? = "",
     val overview: String? = "",
@@ -24,5 +24,5 @@ data class Movie(
     val title: String? = "",
     val video: Boolean? = false,
     val vote_average: Double? = 0.0,
-    val vote_count: Int? = 0
+    val vote_count: Long? = 0
 ) : Parcelable
